@@ -8,54 +8,38 @@
     "license": "MIT",
     "require": {
         "php": "^8.0",
-        "barryvdh/laravel-ide-helper": "^2.6",
-        "codex-team/editor.js": "dev-master",
         "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^2.0",
         "graham-campbell/markdown": "^13.1",
         "guzzlehttp/guzzle": "^7.0.1",
         "itsgoingd/clockwork": "^5.1",
         "jackiedo/dotenv-editor": "^1.2",
-        "jcc/laravel-vote": "dev-master",
-        "laravel/framework": "^8.12",
-        "laravel/passport": "^10.1",
+        "laravel/framework": "^8.75",
+        "laravel/sanctum": "^2.11",
         "laravel/socialite": "^5.1",
+        "laravel/telescope": "^4.6",
         "laravel/tinker": "^2.5",
-        "laravel/ui": "^3.0",
-        "laravelcollective/html": "^6.0",
+        "laravel/ui": "^3.4",
         "mews/purifier": "~3.0",
-        "overtrue/laravel-favorite": "^4.0",
-        "overtrue/laravel-follow": "^3.1",
+        "multicaret/laravel-acquaintances": "^3.5",
         "parsedown/laravel": "^1.1",
         "predis/predis": "^1.1",
         "qcod/laravel-gamify": "^1.0",
-        "rennokki/laravel-eloquent-query-cache": "^3.1",
         "sentry/sentry-laravel": "^2.3",
-        "spatie/laravel-medialibrary": "^9.0.0"
+        "spatie/laravel-medialibrary": "^9.9"
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.5",
         "facade/ignition": "^2.5",
+        "barryvdh/laravel-ide-helper": "^2.6",
         "fakerphp/faker": "^1.9.1",
         "mockery/mockery": "^1.4.2",
         "nunomaduro/collision": "^5.0",
-        "nunomaduro/larastan": "^1.0.1",
+        "nunomaduro/larastan": "^1.0",
         "nunomaduro/phpinsights": "dev-master",
         "phpunit/phpunit": "^9.3.3",
         "psalm/plugin-laravel": "^1.4",
-        "roave/security-advisories": "dev-master",
         "vimeo/psalm": "^4.6"
-    },
-    "config": {
-        "optimize-autoloader": true,
-        "preferred-install": "dist",
-        "sort-packages": true
-    },
-    "extra": {
-        "laravel": {
-            "dont-discover": [
-            ]
-        }
     },
     "autoload": {
         "psr-4": {
@@ -69,8 +53,6 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
     "scripts": {
         "post-autoload-dump": [
             "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
@@ -88,5 +70,22 @@
             "@php artisan ide-helper:meta",
             "@php artisan ide-helper:models --nowrite"
         ]
-    }
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": false,
+            "composer/package-versions-deprecated": false
+        }
+    },
+    "extra": {
+        "laravel": {
+            "dont-discover": [
+            ]
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
