@@ -2,7 +2,6 @@
 
 namespace App\Http\Middleware;
 
-use App;
 use Closure;
 use Exception;
 use Illuminate\Http\Request;
@@ -20,13 +19,9 @@ class Localization
      */
     public function handle($request, Closure $next): mixed
     {
-        $raw_locale = cache('lang');
-        if (in_array($raw_locale, config('app.locales'))) {
-            $locale = $raw_locale;
-        } else {
-            $locale = config('app.locale');
+        if ($request->header('Accept-Language')) {
+            \App::setLocale($request->header('Accept-Language'));
         }
-        App::setLocale($locale);
 
         return $next($request);
     }
