@@ -13,23 +13,16 @@ class HubResource extends JsonResource
     public function toArray($request)
     {
         return [
-            'type' => 'hubs',
-            'id' => (int) $this->id,
-            'attributes' => [
-                'name' => $this->name,
-                'slug' => $this->slug,
-                'logo' => ($this->logo) ? '/'.strtolower($this->logo) : '/images/empty/code.png',
-                'rating' => $this->rating,
-                //                'description'         => $this->description['az'],
-                'hub_followers_count' => $this->favorites_count > 0 ?
-                    Numeric::number_format_short($this->favorites_count) : '0',
-                'follower_check' => auth()->guard('api')->id() ? $this->isFavoritedBy(auth()->guard('api')->user())
-                    : false,
-                'articles_count' => $this->articles_count,
-            ],
-            'links' => [
-                'self' => route('hubs.api.show', ['hub' => $this->id]),
-            ],
+            'id' => (int)$this->id,
+            'name' => $this->name,
+            'slug' => $this->slug,
+            'logo' => ($this->logo) ? '/'.strtolower($this->logo) : '/images/empty/code.png',
+            'rating' => $this->rating,
+            //                'description'         => $this->description['az'],
+            'hub_followers_count' => Numeric::number_format_short($this->followers_count),
+            'follower_check' => auth('sanctum')->id() ? $this->isFollowedBy(auth('sanctum')->user())
+                : false,
+            'articles_count' => $this->articles_count
         ];
     }
 }
