@@ -7,8 +7,8 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
-use Overtrue\LaravelFavorite\Traits\Favoriteable;
-use Rennokki\QueryCache\Traits\QueryCacheable;
+use Multicaret\Acquaintances\Traits\CanBeFollowed;
+use Multicaret\Acquaintances\Traits\CanBeVoted;
 
 /**
  * Class Hub.
@@ -17,17 +17,9 @@ use Rennokki\QueryCache\Traits\QueryCacheable;
  */
 final class Hub extends Model
 {
+    use CanBeFollowed;
+    use CanBeVoted;
     use HasFactory;
-    use Favoriteable;
-    use QueryCacheable;
-
-    public $cacheFor = 3600;
-
-    public $cacheTags = ['hubs'];
-
-    public $cachePrefix = 'hubs_';
-
-    protected static $flushCacheOnUpdate = true;
 
     public $timestamps = false;
 
