@@ -15,12 +15,7 @@ return [
     |
     */
 
-    'paths' => [
-        'api/*',
-        '/login',
-        '/logout',
-        'sanctum/csrf-cookie',
-    ],
+    'paths' => ['*'],
 
     'allowed_methods' => ['*'],
 
