@@ -3,40 +3,44 @@
 @section('title')Login səhifəsi @stop
 
 @section('main')
-    <div class="flex">
-        {{--        <div class="login__left">--}}
-        {{--            --}}{{--                @include('auth.import.left_side')--}}
-        {{--            --}}{{--                <div id="particles-js"></div>--}}
-        {{--        </div>--}}
-        <div class="xs:pt-20 text-center absolute top-1/2 left-1/2 -mr-1/2" style="transform: translate(-50%, -50%)">
-            <a href="{{ route('home') }}"
-               class="text-black font-bold hover:text-gray-200 md:text-2xl m-auto">
-                <img src="{{ asset('images/DevHub_Monochrome_Full_Logo.svg') }}" alt="DevHub Logo" width="188"
-                     height="36" class="mb-4 mx-auto">
-            </a>
-            <login></login>
+    <div class="container">
+        <div class="row justify-content-center">
+            <div class="col-md-8">
+                <div class="card">
+                    <div class="card-header">{{ __('Reset Password') }}</div>
 
-            <div class="w-full border-t border-gray-300 my-4"></div>
-            <div class="flex justify-between">
-                <div class="flex items-center">
-                    <i class="iconify inline-block m-auto w-10 h-10 p-2 bg-gray-700 rounded-lg text-white" data-icon="mdi-thumbs-up-down"></i>
-                    <div class="text-left text-sm ml-2">
-                        <p>Ses</p>
-                        <p>ver</p>
-                    </div>
-                </div>
-                <div class="flex">
-                    <i class="iconify inline-block m-auto w-10 h-10 p-2 bg-gray-700 rounded-lg text-xl text-white" data-icon="mdi-layers-outline"></i>
-                    <div class="text-left text-sm ml-2">
-                        <p>Paylaşmalar</p>
-                        <p>yaz</p>
-                    </div>
-                </div>
-                <div class="flex">
-                    <i class="iconify inline-block m-auto w-10 h-10 p-2 bg-gray-700 rounded-lg text-xl text-white" data-icon="mdi-bookmark-multiple-outline"></i>
-                    <div class="text-left text-sm ml-2">
-                        <p>Izləici</p>
-                        <p>ol</p>
+                    <div class="card-body">
+                        @if (session('status'))
+                            <div class="alert alert-success">
+                                {{ session('status') }}
+                            </div>
+                        @endif
+
+                        <form method="POST" action="{{ route('password.email') }}">
+                            @csrf
+
+                            <div class="form-group row">
+                                <label for="email" class="col-md-4 col-form-label text-md-right">{{ __('E-Mail Address') }}</label>
+
+                                <div class="col-md-6">
+                                    <input id="email" type="email" class="form-control{{ $errors->has('email') ? ' is-invalid' : '' }}" name="email" value="{{ old('email') }}" required>
+
+                                    @if ($errors->has('email'))
+                                        <span class="invalid-feedback">
+                                        <strong>{{ $errors->first('email') }}</strong>
+                                    </span>
+                                    @endif
+                                </div>
+                            </div>
+
+                            <div class="form-group row mb-0">
+                                <div class="col-md-6 offset-md-4">
+                                    <button type="submit" class="btn btn-primary">
+                                        {{ __('Send Password Reset Link') }}
+                                    </button>
+                                </div>
+                            </div>
+                        </form>
                     </div>
                 </div>
             </div>
